Move proptype-validator utility into src
Only src files are transpiled for external use, so requiring this file was breaking my jest tests
