+++ packages/components/src/card/index.js
@@ -11,7 +11,7 @@ import PropTypes from 'prop-types';
  */
 import EllipsisMenu from '../ellipsis-menu';
 import { H, Section } from '../section';
-import { validateComponent } from '../../lib/proptype-validator';
+import { validateComponent } from '../lib/proptype-validator';
 
 /**
  * A basic card component with a header. The header can contain a title, an action, and an `EllipsisMenu` menu.
