+++ packages/components/src/lib/proptype-validator.js
